ödendi iade butonları eklendi

+++ app/Http/Controllers/SiparisController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 
 
+use App\Models\Kasadefteri;
 use App\Models\Musteri;
 use App\Models\Siparis;
 use Illuminate\Http\Request;
@@ -22,12 +23,12 @@ class SiparisController extends Controller
         $musteri = Musteri::find($musteri_id);
         $siparisler = DB::table('siparises')->where('musteri_id', $musteri_id)->get();
         if (count($siparisler) == 0) {
-            return view('treetoner.siparisler.index', ['musteri_id' => $musteri_id,'musteri'=>$musteri]);
+            return view('treetoner.siparisler.index', ['musteri_id' => $musteri_id, 'musteri' => $musteri]);
         } else {
             return view('treetoner.siparisler.index', [
-             'siparisler' => $siparisler,
-             'musteri_id' => $musteri_id,
-             'musteri'=>$musteri
+                'siparisler' => $siparisler,
+                'musteri_id' => $musteri_id,
+                'musteri' => $musteri
             ]);
         }
 
@@ -75,6 +76,7 @@ class SiparisController extends Controller
         $siparis->sonuc = $request->get('sonuc');
         $siparis->fiyat = $request->get('fiyat');
         $siparis->musteri_id = $request->get('musteri_id');
+        $siparis->tahsilat = "money_wait";
         $siparis->save();
 
 
@@ -107,25 +109,84 @@ class SiparisController extends Controller
             $siparis->power_kablo = "yok";
 
         }
-        $siparis->save();
+       
+        if ($siparis->tahsilat == "money_paid") {
+           
 
-        return redirect()->route('siparis.index', ['musteri_id' => $siparis->musteri_id]);
+            return redirect()->route('siparis.index', ['musteri_id' => $siparis->musteri_id])->with('fail','sipariş kasa defterine işlendiğinden güncelleme yapılamaz.');
+
+        }else{
+            $siparis->save();
+
+            return redirect()->route('siparis.index', ['musteri_id' => $siparis->musteri_id]);
+        }
+      
     }
     public function destroy($siparis_id, $musteri_id)
     {
+
         $siparis = Siparis::find($siparis_id);
-        $siparis->delete();
-        // For a route with the following URI: profile/{id}
+        if ($siparis->tahsilat == "money_paid") {
+            return redirect()->route('siparis.index', ['musteri_id' => $musteri_id])->with('fail','kasadefterine kaydı yapıldığı için siparişi silemezsiniz.');
 
-        return redirect()->route('siparis.index', ['musteri_id' => $musteri_id]);
-        
+        }
+        if ($siparis->tahsilat == "money_return") {
+            $siparis->delete();
+            return redirect()->route('siparis.index', ['musteri_id' => $musteri_id]);
+        }
+        if ($siparis->tahsilat == "money_wait") {
+            $siparis->delete();
+            return redirect()->route('siparis.index', ['musteri_id' => $musteri_id]);
+        }
+      
+
+
     }
-    public function siparis_goster($siparis_id){
+    public function siparis_goster($siparis_id)
+    {
         $siparis = Siparis::find($siparis_id);
-        return view('treetoner.siparisler.siparis_goster',['siparis'=>$siparis]);
-        
+        return view('treetoner.siparisler.siparis_goster', ['siparis' => $siparis]);
+
 
 
     }
+
+    public function tahsilatlar()
+    {
+        $siparisler = Siparis::where('tahsilat', '!=', 'money_paid')->get();
+        return view('treetoner.tahsilatlar.index', ['siparisler' => $siparisler]);
+    }
+    public function tahsilat_money_paid($siparis_id)
+    {
+        $siparis = Siparis::find($siparis_id);
+        $siparis->tahsilat = "money_paid";
+        $siparis->save();
+        //ödemesi yapılan tahsilat kasa defterine atılacak...
+        $gelir = new Kasadefteri;
+        $musteri = Musteri::find($siparis->musteri_id);
+        if ($siparis->yazici_model == "") {
+            $gelir->aciklama = $musteri->adi_soyadi . " isimli müşterinin Yazıcı Tamir/Bakım ücreti (" . $siparis->yazıcı_model . ")";
+        } else {
+            $gelir->aciklama = $musteri->adi_soyadi . " isimli müşterinin Yazıcı Tamir/Bakım ücreti";
+        }
+        $gelir->fiyat = $siparis->fiyat;
+        $gelir->save();
+        $siparisler = Siparis::where('tahsilat', '!=', 'money_paid')->get();
+        return view('treetoner.tahsilatlar.index', ['siparisler' => $siparisler]);
+
+
+    }
+    public function tahsilat_money_return($siparis_id)
+    {
+        $siparis = Siparis::find($siparis_id);
+        $siparis->tahsilat = "money_return";
+        $siparis->save();
+
+        $siparisler = Siparis::where('tahsilat', '!=', 'money_paid')->get();
+        return view('treetoner.tahsilatlar.index', ['siparisler' => $siparisler]);
+
+
+    }
+
 
 }
